dev-tools/epub-bar-chart-solr-relevancy-scores/: clicking results header toggles between chart types -- all pages chart not vetted yet [+]
Originally toggling was done by clicking on the bar chart, but if user clicked while a tooltip was visible, it stayed there.
This is due to pre-existing bug which caused a new tooltip <div> to be created every time drawChart is called, which
happens when new EPUB is selected, and now also happens when bar chart type is toggled.